requirements.txt
discord.py[voice]
requests
python-dotenv
PyNaCl
yt-dlp
psutil
